+++ upload-api/src/Command/GetRemoteUrlsCommand.php
@@ -15,6 +15,7 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\PropertyAccess\Exception\UninitializedPropertyException;
 
 #[AsCommand(
     name: 'app:cover:get-remote-urls',
@@ -54,15 +55,19 @@ class GetRemoteUrlsCommand extends Command
 
         /** @var Cover $cover */
         foreach ($query->toIterable() as $cover) {
-            if ($this->coverStoreService->exists($cover->getMaterial()->getIsIdentifier())) {
-                $this->coverStoreService->removeLocalFile($cover);
-                $item = $this->coverStoreService->search($cover->getMaterial()->getIsIdentifier());
+            try {
+                if ($this->coverStoreService->exists($cover->getMaterial()->getIsIdentifier())) {
+                    $this->coverStoreService->removeLocalFile($cover);
+                    $item = $this->coverStoreService->search($cover->getMaterial()->getIsIdentifier());
 
-                if (null !== $item) {
-                    $cover->setRemoteUrl($item->getUrl());
+                    if (null !== $item) {
+                        $cover->setRemoteUrl($item->getUrl());
+                    }
+                    $cover->setUploaded(true);
+                    $this->entityManager->flush();
                 }
-                $cover->setUploaded(true);
-                $this->entityManager->flush();
+            } catch (UninitializedPropertyException $e) {
+                // Do nothing, as cover was not connect to a material.
             }
         }